provisioner/requirements.txt
boto3==1.4
botocore==1.7.45
PyYAML==3.12

